requirements.txt
Flask
PyPDF2
pillow
pytesseract
transformers
torch
sentencepiece
werkzeug
